move files structures
